+++ examples/test/test_pwscf_scf.py
@@ -3,18 +3,19 @@ import qepy
 import unittest
 import pathlib
 import shutil
+try:
+    from mpi4py import MPI
+    comm = MPI.COMM_WORLD
+    commf = comm.py2f()
+except Exception:
+    comm = None
+    commf = None
 
 class Test(unittest.TestCase):
     def test_scf(self):
-        try:
-            from mpi4py import MPI
-            comm = MPI.COMM_WORLD
-            comm = comm.py2f()
-        except Exception:
-            comm = None
         path = pathlib.Path(__file__).resolve().parent / 'DATA'
         fname = path / 'qe_in.in'
-        qepy.qepy_pwscf(fname, comm)
+        qepy.qepy_pwscf(fname, commf)
         embed = qepy.qepy_common.embed_base()
         # embed.ldescf = True # add scf correction energy
         embed.iterative = True
@@ -34,12 +35,13 @@ class Test(unittest.TestCase):
         qepy.qepy_stop_run(0, what = 'no')
 
     def tearDown(self):
-        path = pathlib.Path('.')
-        for f in path.glob('al.*'):
-            if f.is_file():
-                f.unlink()
-            else :
-                shutil.rmtree(f)
+        if comm and comm.rank == 0 :
+            path = pathlib.Path('.')
+            for f in path.glob('al.*'):
+                if f.is_file():
+                    f.unlink()
+                else :
+                    shutil.rmtree(f)
 
 
 if __name__ == "__main__":
